rpl: Implement case statement
Implement the RPL `case` statement, including a `case...when`
extension to traditional RPL.

The `case` statement can be used to select one case among many.
Inside a `case`, there is a list of conditions, each followed by
`then` or `when`. Code following `then` or `when` is executed when the
condition is met.

* A condition preceding `then` is a boolean condition, similar to the
  condition in an `if` statement.

* A condition preceding `when` is a value that must match the current
  value on the stack exactly.

For example, the following tests the sign of `X`:

```
X
case
  dup 0 < then "Neg" end
  dup 0 > then "Pos" end
  "Zero"
end
```

The `when` syntax is useful to test exact values, for example, you can
check return the spelling of the ten digits with:

```
X
case
   0 when "zero"  end
   1 when "one"   end
   2 when "two"   end
   3 when "three" end
   4 when "four"  end
   5 when "five"  end
   6 when "size"  end
   7 when "seven" end
   8 when "eight" end
   9 when "nine"  end
end
```

Fixes: #374

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>