rpl: Call reset_auto_off regularly
When `reset_auto_off()` is not called regularly, the keyboard stops
responding after a few minutes of execution. The RPL program keeps executing,
but cannot be interrupted and has to run until completion.

Adding `reset_auto_off` in `program::interrupted` allows the program to be
interrupted after 10 minutes of execution. Tested on the new `Patterns` demo.

Fixes: #587

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>