beep: Implement `Beep` command
Implement the `Beep` command.

The tricky part was making sure that beeps could be interrupted,
while at the same time allowing single-stepping of `Beep`.

The HP documentation allows beep to last up to 1200s.
That seems quite a lot for the little battery in the DM42/DB42.
Decided to max out at 10s instead.

Fixes: #50

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>