ifte: Execute program arguments
As pointed out by the RPL article on Wikipedia, `IFT` and `IFTE`
execute program arguments. The given example is:

```
«
   DUP 1 ==
   « DROP "One" »
   « 1 < "Less" "More" IFTE »
   IFTE
 »
 ```

Push the program body on the run stack to ensure that we execute it.

Fixes: #592

Also factor out conditional evaluation of the condition in `if`,
`while` and `case`, so that we evaluate programs and expressions.

Note that the evaluation of programs in that context is a deviation
from standard HP calculators, which only evaluate expressions.

Fixes: #593

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>