wait: Allow `Wait` command to be interrupted by `EXIT` key
When using the `Wait` command, a program became harder to interrupt,
because `Wait` itself was consuming the keys. Ensure that if we get
the Exit key, we interrupt the program. We still return the key to
allow `Continue` to do its job correctly if the program is expecting
something on the stack.

Also add a `power_check` call inside `Wait`, a function carved out
from the main loop, and use a timer instead of `sys_delay` inside the
`Wait` command, with two benefits:

* Lower power usage for long waits. Instead of waking up 20 times per
  second to check if there was a key, using a timer allows us to sleep
  almost forever. By waiting 1 minute at a time, we wake up at most 10
  times for a very short duration before the auto-off kicks in.

* `0 WAIT` and `-1 WAIT` will respect the auto-off. The calculator
  powers off  with a message indicating it went into power saving mode
  instead of drawing the usual off image. But then it is really off,
  meaning it consumes practically no power whatsoever. And whatever
  your program had been drawing is still largely on the screen.

The `program::interrupted` code was moved from `main.cc` to
`program.cc`, which is a more logical place for it.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>