Add commands to build text to and from Unicode
`Char→Code` (`NUM`, `CODEPOINT`, `CharToUnicode`) returns the Unicode
codepoint of the first character in the text, or `-1` if the text is
empty. `"Hello" NUM` returns `72`.

`Text→Code` (`TextToUnicode`) returns a list of the Unicode codepoints
for all codepoints in the given text.  `"Hello" Text→Code` returns `{
72 101 108 108 111 }`.

`Code→Text` (`CHR`, `Code→Char`) builds a text out of a Unicode
codepoint. The argument can be either a single numerical value or a
list of numerical values. A negative numerical value produces an empty
text. `{ 87 111 114 -22 108 100 }` returns `"World"`, the value `-22`
producing no character. `42 CHR` returns `"*"`, and `34 CHR` returns
`""""`, which is a 1-character text containing an ASCII quote `"`.

Fixes: #597

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>