Makefile: Add `check-ids` target
The `check-ids` target checks if commands are in the menus
or in the help.

The `src/ignored_menus.csv` and `src/ignored_help.csv` indicate what
does not need to be reported by the tools. It was initialized with the
current state of missing things, and can serve as a reference for what
commands need to either be documented or added to menus.

Fixes: #615

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>