list: Implement `map`, `reduce` and `filter`
`Map` applies an operation on all elements in a list or array. The
operation on the first level of the stack should take one argument and
return a single value.

`Reduce` applies a cumulative pairwise operation on all elements in a
list or array.  The operation on the first level of the stack should
take two arguments and combine them into a single value. The result is
the repeated application of that operation to all elements.

`Filter` selects elements in a list of array based on a predicate. The
predicate given on level 1 of the stack takes a value as argument, and
returns a truth values. The resulting list is built with all elements
where the predicate is true.

Fixes: #613

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>