commands: Return HP-compatible values from `Type`
The `Type` command is used in many programs. It is useful to have it
return HP-compatible values, even if the mapping is not entirely
accurate.

This is under control of the `DetailedTypes` / `CompatibleTypes`
settings. When in `CompatibleTypes` mode, an attempt is made to return
a compatible value, which lacks details. When in `DetailedTypes` mode,
the value returned is negative (to distinguish it from HP values) and
is gives precise information about the actual internal representation.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>