flags: Implement compatibility code for negative flags
Recognize values such as `-64` as input for `CF` or `SF`.

This allows code that uses the operational flags to work correctly,
notably the "wrap around" flag (-64) set by `GetI` and `SetI`, as in
the `MNX` and `MNX2` exxamples found on page 2-16 and 2-17 of the
HP50G advanced reference manual. Note that if writing code
specifically for DB48X, it's probably better to use the symbolic name
for the flag, i.e. `IndexWrapped` in that case.

Fixes: #625

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>