catalog: Auto-complete all possible spellings
Allow the catalog to auto-complete all possible command spellings.
As a side effect, accelerate the parsing of commands.

All spellings are now stored in a static array, `object::spellings`.
This makes it possible to scan the table in both directions, i.e.
from `id` to text for name rendering, or from name to `id` for
parsing. This makes it also simpler to run operations that loop on all
spellings, including building the catalog entries.

Fixes: #626

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>