arithmetic: Based number promotion
Align the behaviour when mixing based numbers and integers to
what the HP calculators do... with deliberate variations.

* Do not accept mixed based / decimal operations
* Convert negative values correctly.

The latter is arguably a bug fix. On HP calculators, `#F 3 -` gives
`#C`, which is the expected value. But `#F -3 +` gives `#F`, which is
a bit curious. DB48X will return `#C` in that case as well. It does so
by converting negative values to equivalent binary representation.

Fixes: #629

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>