solver: Change default settings for solver
The default settings for the solver cause it to not find root for simple
equations. This is primarily because the value for `Precision` required all
digits to be exact, which does not work for really simple equations.

It did not help that the default value for the number of iterations was 10,
which is much too small to solve things when you start near a minimum.

Fixes: #627

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>