settings: Accept large `Fix`, `Sci`, `Eng`
With variable-precision decimals, we may want to have a precision like
99 and therefore might want to accept `99 FIX`.

Also select a default precision of 24 digits. Precisions that are
multiples of 12 make optimal use of the available bits, because they
align both on 10-bit "kigit" and 8-bit "byte" boundaries.

A 12-digit representation uses up to 5 bytes for digits, 1 byte of ID,
1 bytes for the precision, and at least one byte for the exponent.
That's 8 bytes for exponents between -64 and 63. That could be a good
default, also matching the HP48 built-in precision.

A 24-bit representation uses up to 10 bytes of payload, so 13 or 14
bytes depending on the exponent. That seems like a good choice for the
default setting, and highlights that the precision is higher than the
HP48. It is also using a little less than the BID128 on the DM42, and
might be a bit faster.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>