decimal: Implement sine, cosine and tangent functions
We use a relatively bare-bones implementation based on Taylor series.
The benefit of that approach is that it only requires one constant,
`pi`, whereas more efficient methods require pre-computed tables.

The pre-computed tables would typically grow as O(n^2) with respect to
precision. This is because you need in the order of n constants and
each of them is in the order of n digits of precision. There are smart
ways to compress things a bit (this is the approach taken in newRPL),
but my approach at the moment is to be as conservative as possible
with respect to code and data space in order to fit on a DM42.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>