leb128: Fix case of signed leb128 value 64
A signed value like 64 would stop encoding at byte 0, resulting in
encoding 0xC0, which would then be interpreted as a negative value.
The correct encoding in that case is 0x40 0x80, which indicates that
we have a zero sign.

This was manifested by the more heavy use of signed leb128 in
variable-precision decimals.

Fixes: #642

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>