command: Do not parse `IfThenElse` as a 1-byte command
A few control structures like `IfThenElse` or `DoWhile`  were moved in
the 1-byte range to minimize memory usage.

However, they are complex commands, and their size is not one byte.
Only the type ID is 1 byte, but there is a body.

Unfortunately, there is also a command that has `IfThenElse` as an
alias. We should not parse `IfThenElse` as the 1-byte prefix, but as
the `IfThenElse` command.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>