decimal: Implement true decimal-to-integer conversion
There was no real decimal to integer conversion, so we were
relying on a fraction approximation.

Add code to perform conversion to integer and bignum.

The original code for `from_bignum` was using  an incremental scale
that was a power of 2 and not powers of 10. This was causing rounding
errors for very large numbers, e.g. `1.234567E67` would convert
correctly to an integer, but the conversion back to `decimal` would
not end with zeroes but with nines. For that reason, the code was
changed into an approach that does exact decimal computations on the
bignum, which is slightly more expensive, since it incurs a
division. We use one divison per group of 12 digits to keep things
relatively fast.

Fixes: #648

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>