unit: Rendering of `dms` and `hms` units
Render `1_dms` unit as `1°00′00″`
Render `1_hms` unit as `1:00:00`

In both cases, the internal representation remains unchanged,
i.e. it can be a decimal or a fraction for exact computations.

When edting, the unit is displayed normally, i.e. `1_dms`.

This exposed a number of formatting issues with decimal rendering,
see #652 and #653.

Fixes: #650

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>