runloop: Store `end` pointer to be "inside" object
In the run loop, the end pointer was not inside an object, but just
one past the object (i.e. it was the result of a program's `skip()`)
As a result, if that `end` pointer was equal to `rt.Temporaries`,
i.e. the start of the free space, and a garbage collection occured,
the start of the program would move but the end would not (it was seen
as not requiring garbage collection).

Fixes: #679

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>