ui: Avoid overlap between busy and ABC indicators
The location of the busy cursor was incorrectly computed, and as a
result would end up erasing part of the ABC indicator.

Additionally, the background of the ABC indicator was not always
erased, so that `abc` would occasionally be drawn on top of `ABC`
instead of on top of a black background.

Fixes: #680

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>