complex: Save `this` in a GC pointer when it can move
There are a few places where we do several computations, e.g. compute
`re` and `im`, which may cause a garbage collection, invalidating the
`this` pointer.

Therefore, it is not legit to write something like `re()+im()`.
It must be replaced with something like `o.re() + o.im()`, where `o`
is a garbage-collected pointer.

Fixes: #745

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>