Allow `·` as a multiplication sign in equations
We are currently using the `×` symbol for multiplication, or `*` in HP48
compatibility mode.

This makes sense for rendering in programs. However, inside expressions, the `·`
symbol is probably better, in particular if the common name `x` is used in the
equation. In other words, `x·y` is more readable than `x×y`.

This is controled by a user preference, `UseDotForMultiplication`, which is
enabled by default. `UseCrossForMultiplication` is the opposite settings, which
restores the previous behaviour.

Fixes: #765

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>