date/time: Refactor the code and add Julian day number
Refactor the code to put all date and time related functions into the
`datetime.h` and `datetime.cc` files.

Also added Julian day number calculations, which enable the
computation of the day of the week in the date format rendering, and
will make it easier to implement future commands such as `DDAYS`.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>