arithmetic: Symbolic and numeric infinity
Add support for the `InfinityValue` (-22) flag, described in the
HP50G as the "Infinite result exception" flag.

When the flag is set, infinite results are returned either as symbolic
constant infinity `∞` or as its value, currently defined as
`9.99999E999999`.

Whether the infinity value is returned symbolically or numerially
depends on the `NumericalConstants` (-2) and `NumericalResults` (-3)
flags.

Fixes: #835

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>