simulator: Add screenshot keys
When hitting the F9 key, take a partial screen snapshot directly
(partial meaning no header).

When hitting the F8 key, take a screen snapshot using the `Screenshot`
command, which is now implemented on the simulator.

Also reduce the size of the window a little so that we can take video
snapshots that are 720 x 400 pixels and capture the keyboard.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>