ui: Insert commands for unit conversions, constants and variables
When typing a program, insert commands for unit conversion, constants
and variable menu entries.

For variables, it used to work, but the way this was happening was
fragile, and had been broken by changes in what commands were seen as
"immediate".

Add tests for all these cases to make sure we catch future breakage.

Also emit an error if trying to insert `ConvertToUnitPrefix in a
program.

Fixes: #848

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>