save: Remove `raw`
The selection of the format for saving is now done in `sysmenu.c`

This is only for standard (startup) mode.
For `STO` to a file, it makes sense to respect user preferences.
But that means you can only reload with the same settings, and
must avoid separators that you cannot parse back.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>