unitfile: Avoid the slight risk of 'what' moving
When we are reading multiple columns in a unit file, it's possible, if
unlikley, that a garbage collection might occur due to the allocation
of symbol in the first column, leading to corrupt data for the `what`
variable.

Note that in the current iteration of the code, I do not think that
"what" itself is sensible once we have set "def", but better code
defensively that hit subtle garbage-collection issues later.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>