catalog: Fix a subtle bug on DM32 with the catalog content
With the original code, the loop searching for commands to add to the
catalog skips many commands. I suspect this is another case of QSPI
hammering similar to what was observed for fonts. In any case, the
workaround appears to be to slightly deoptimize the code by selecting
a lower optimization level and adding an `else` clause in the inner
`if` statement.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>